+++ docs/source/example-code/src/test/kotlin/net/corda/docs/CustomVaultQueryTest.kt
@@ -1,7 +1,7 @@
 package net.corda.docs
 
 import net.corda.core.contracts.Amount
-import net.corda.core.node.services.ServiceInfo
+import net.corda.core.identity.Party
 import net.corda.core.utilities.OpaqueBytes
 import net.corda.core.utilities.getOrThrow
 import net.corda.finance.*
@@ -9,11 +9,13 @@ import net.corda.finance.contracts.getCashBalances
 import net.corda.finance.flows.CashIssueFlow
 import net.corda.node.internal.StartedNode
 import net.corda.finance.schemas.CashSchemaV1
+import net.corda.nodeapi.ServiceInfo
 import net.corda.node.services.network.NetworkMapService
 import net.corda.node.services.transactions.ValidatingNotaryService
 import net.corda.testing.DUMMY_NOTARY
 import net.corda.testing.DUMMY_NOTARY_KEY
 import net.corda.testing.chooseIdentity
+import net.corda.testing.getDefaultNotary
 import net.corda.testing.node.MockNetwork
 import org.junit.After
 import org.junit.Assert
@@ -27,6 +29,7 @@ class CustomVaultQueryTest {
     lateinit var notaryNode: StartedNode<MockNetwork.MockNode>
     lateinit var nodeA: StartedNode<MockNetwork.MockNode>
     lateinit var nodeB: StartedNode<MockNetwork.MockNode>
+    lateinit var notary: Party
 
     @Before
     fun setup() {
@@ -43,6 +46,7 @@ class CustomVaultQueryTest {
         nodeA.internals.installCordaService(CustomVaultQuery.Service::class.java)
         nodeA.internals.registerCustomSchemas(setOf(CashSchemaV1))
         nodeB.internals.registerCustomSchemas(setOf(CashSchemaV1))
+        notary = nodeA.services.getDefaultNotary()
     }
 
     @After
@@ -71,7 +75,7 @@ class CustomVaultQueryTest {
         // Use NodeA as issuer and create some dollars
         val flowHandle1 = nodeA.services.startFlow(CashIssueFlow(amountToIssue,
                 OpaqueBytes.of(0x01),
-                notaryNode.info.notaryIdentity))
+                notary))
         // Wait for the flow to stop and print
         flowHandle1.resultFuture.getOrThrow()
     }
@@ -81,7 +85,7 @@ class CustomVaultQueryTest {
                 nodeA.info.chooseIdentity(),
                 OpaqueBytes.of(0x01),
                 nodeA.info.chooseIdentity(),
-                notaryNode.info.notaryIdentity))
+                notary))
         flowHandle1.resultFuture.getOrThrow()
     }