+++ node/src/main/kotlin/net/corda/node/services/CoreFlowHandlers.kt
@@ -38,7 +38,7 @@ class NotaryChangeHandler(otherSide: Party) : AbstractStateReplacementFlow.Accep
         }
 
         // TODO: load and compare against notary whitelist from config. Remove the check below
-        val isNotary = serviceHub.networkMapCache.notaryNodes.any { it.notaryIdentity == newNotary }
+        val isNotary = serviceHub.networkMapCache.isNotary(newNotary)
         if (!isNotary) {
             throw StateReplacementException("The proposed node $newNotary does not run a Notary service")
         }