+++ tools/loadtest/src/main/kotlin/net/corda/loadtest/Disruption.kt
@@ -42,9 +42,10 @@ data class DisruptionSpec(
  *     * Randomly block queues.
  *     * Randomly duplicate messages, perhaps to other queues even.
  */
-
-val isNetworkMap = { node: NodeConnection -> node.info.advertisedServices.any { it.info.type == NetworkMapService.type } }
-val isNotary = { node: NodeConnection -> node.info.advertisedServices.any { it.info.type.isNotary() } }
+val isNotary = { node: NodeConnection ->
+    val notaries = node.proxy.notaryIdentities()
+    node.info.legalIdentitiesAndCerts.any { it in notaries }
+}
 fun <A> ((A) -> Boolean).or(other: (A) -> Boolean): (A) -> Boolean = { this(it) || other(it) }
 
 fun hang(hangIntervalRange: LongRange) = Disruption("Hang randomly") { node, random ->