+++ docs/source/changelog.rst
@@ -6,6 +6,19 @@ from the previous milestone release.
 
 UNRELEASED
 ----------
+* ``NodeInfo`` and ``NetworkMapCache`` changes:
+   * Removed ``NodeInfo::legalIdentity`` in preparation for handling of multiple identities. We left list of ``NodeInfo::legalIdentitiesAndCerts``,
+   the first identity still plays a special role of main node identity.
+   * We no longer support advertising services in network map. Removed ``NodeInfo::advertisedServices``, ``serviceIdentities``
+   and ``notaryIdentity``.
+   * Removed service methods from ``NetworkMapCache``: ``partyNodes``, ``networkMapNodes``, ``notaryNodes``, ``regulatorNodes``,
+   ``getNodesWithService``, ``getPeersWithService``, ``getRecommended``, ``getNodesByAdvertisedServiceIdentityKey``,
+   ``notaryNode``, ``getAnyServiceOfType``. To get all known ``NodeInfo``s call ``allNodes``.
+   * In preparation for ``NetworkMapService`` redesign and distributing notaries through ``NetworkParameters`` we added
+   ``NetworkMapCache::notaryIdentities`` list to enable to lookup for notary parties known to the network. Related ``CordaRPCOps::notaryIdentities``
+   was introduced. Other special nodes parties like Oracles or Regulators need to be specified directly in CorDapp or flow.
+   * Moved ``ServiceType`` and ``ServiceInfo`` to ``net.corda.nodeapi`` package as services are only required on node startup.
+
 * Adding enum support to the class carpenter
 
 * ``ContractState::contract`` has been moved ``TransactionState::contract`` and it's type has changed to ``String`` in order to
@@ -153,6 +166,40 @@ UNRELEASED
 * ``X509CertificateHolder`` has been removed from the public API, replaced by ``java.security.X509Certificate``.
 
 * Moved ``CityDatabase`` out of ``core`` and into ``finance``
+
+* All of the ``serializedHash`` and ``computeNonce`` functions have been removed from ``MerkleTransaction``.
+  The ``serializedHash(x: T)`` and ``computeNonce`` were moved to ``CryptoUtils``.
+
+* Two overloaded methods ``componentHash(opaqueBytes: OpaqueBytes, privacySalt: PrivacySalt, componentGroupIndex: Int,
+  internalIndex: Int): SecureHash`` and ``componentHash(nonce: SecureHash, opaqueBytes: OpaqueBytes): SecureHash`` have
+  been added to ``CryptoUtils``. Similarly to ``computeNonce``, they internally use SHA256d for nonce and leaf hash
+  computations.
+
+* The ``verify(node: PartialTree, usedHashes: MutableList<SecureHash>): SecureHash`` in ``PartialMerkleTree`` has been
+  renamed to ``rootAndUsedHashes`` and is now public, as it is required in the verify function of ``FilteredTransaction``.
+
+* ``TraversableTransaction`` is now an abstract class extending ``CoreTransaction``. ``WireTransaction`` and
+  ``FilteredTransaction`` now extend ``TraversableTransaction``.
+
+* Two classes, ``ComponentGroup(open val groupIndex: Int, open val components: List<OpaqueBytes>)`` and
+  ``FilteredComponentGroup(override val groupIndex: Int, override val components: List<OpaqueBytes>,
+      val nonces: List<SecureHash>, val partialMerkleTree: PartialMerkleTree): ComponentGroup(groupIndex, components)``
+  have been added, which are properties of the ``WireTransaction`` and ``FilteredTransaction``, respectively.
+
+* ``checkAllComponentsVisible(componentGroupEnum: ComponentGroupEnum)`` is added to ``FilteredTransaction``, a new
+  function to check if all components are visible in a specific component-group.
+
+* To allow for backwards compatibility, ``WireTransaction`` and ``FilteredTransaction`` have new fields and
+  constructors: ``WireTransaction(componentGroups: List<ComponentGroup>, privacySalt: PrivacySalt = PrivacySalt())``,
+  ``FilteredTransaction private constructor(id: SecureHash,filteredComponentGroups: List<FilteredComponentGroup>,
+      groupHashes: List<SecureHash>``. ``FilteredTransaction`` is still built via
+  ``buildFilteredTransaction(wtx: WireTransaction, filtering: Predicate<Any>).
+
+* ``FilteredLeaves`` class have been removed and as a result we can directly call the components from
+  ``FilteredTransaction``, such as ``ftx.inputs`` Vs the old ``ftx.filteredLeaves.inputs``.
+
+* A new ``ComponentGroupEnum`` is added with the following enum items: ``INPUTS_GROUP``, ``OUTPUTS_GROUP``,
+ ``COMMANDS_GROUP``, ``ATTACHMENTS_GROUP``, ``NOTARY_GROUP``, ``TIMEWINDOW_GROUP``.
 
 Milestone 14
 ------------